api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
frontend==0.0.3
pymupdf==1.23.7
pandas==2.1.4
pytesseract==0.3.10
python-dotenv==1.0.0
openpyxl==3.1.2
openai==1.3.9
tiktoken==0.5.2
faiss-cpu==1.7.4